// src/lib/api/generated/models/v1AccountStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1AccountStatus {
    ACCOUNT_STATUS_UNSPECIFIED = 'ACCOUNT_STATUS_UNSPECIFIED',
    ACCOUNT_STATUS_ACTIVE = 'ACCOUNT_STATUS_ACTIVE',
    ACCOUNT_STATUS_INACTIVE = 'ACCOUNT_STATUS_INACTIVE',
}