// src/lib/api/generated/models/v1CardBrand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1CardBrand {
    CARD_BRAND_UNSPECIFIED = 'CARD_BRAND_UNSPECIFIED',
    CARD_BRAND_VISA = 'CARD_BRAND_VISA',
    CARD_BRAND_MASTERCARD = 'CARD_BRAND_MASTERCARD',
    CARD_BRAND_DINACARD = 'CARD_BRAND_DINACARD',
    CARD_BRAND_AMEX = 'CARD_BRAND_AMEX',
}