// src/lib/api/generated/models/v1CardStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1CardStatus {
    CARD_STATUS_UNSPECIFIED = 'CARD_STATUS_UNSPECIFIED',
    CARD_STATUS_ACTIVE = 'CARD_STATUS_ACTIVE',
    CARD_STATUS_BLOCKED = 'CARD_STATUS_BLOCKED',
    CARD_STATUS_DEACTIVATED = 'CARD_STATUS_DEACTIVATED',
}