// src/lib/api/generated/models/v1LoanType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1LoanType {
    LOAN_TYPE_UNSPECIFIED = 'LOAN_TYPE_UNSPECIFIED',
    LOAN_TYPE_CASH = 'LOAN_TYPE_CASH',
    LOAN_TYPE_HOUSING = 'LOAN_TYPE_HOUSING',
    LOAN_TYPE_AUTO = 'LOAN_TYPE_AUTO',
    LOAN_TYPE_REFINANCE = 'LOAN_TYPE_REFINANCE',
    LOAN_TYPE_STUDENT = 'LOAN_TYPE_STUDENT',
}