// src/lib/api/generated/models/v1InterestType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1InterestType {
    INTEREST_TYPE_UNSPECIFIED = 'INTEREST_TYPE_UNSPECIFIED',
    INTEREST_TYPE_FIXED = 'INTEREST_TYPE_FIXED',
    INTEREST_TYPE_VARIABLE = 'INTEREST_TYPE_VARIABLE',
}