// src/lib/api/generated/models/v1LoanStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1LoanStatus {
    LOAN_STATUS_UNSPECIFIED = 'LOAN_STATUS_UNSPECIFIED',
    LOAN_STATUS_APPROVED = 'LOAN_STATUS_APPROVED',
    LOAN_STATUS_REJECTED = 'LOAN_STATUS_REJECTED',
    LOAN_STATUS_PAID_OFF = 'LOAN_STATUS_PAID_OFF',
    LOAN_STATUS_OVERDUE = 'LOAN_STATUS_OVERDUE',
}