// src/lib/api/generated/models/v1InstallmentStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1InstallmentStatus {
    INSTALLMENT_STATUS_UNSPECIFIED = 'INSTALLMENT_STATUS_UNSPECIFIED',
    INSTALLMENT_STATUS_PAID = 'INSTALLMENT_STATUS_PAID',
    INSTALLMENT_STATUS_UNPAID = 'INSTALLMENT_STATUS_UNPAID',
    INSTALLMENT_STATUS_OVERDUE = 'INSTALLMENT_STATUS_OVERDUE',
}