// src/lib/api/generated/models/v1TransactionKind.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1TransactionKind {
    TRANSACTION_KIND_UNSPECIFIED = 'TRANSACTION_KIND_UNSPECIFIED',
    TRANSACTION_KIND_PAYMENT = 'TRANSACTION_KIND_PAYMENT',
    TRANSACTION_KIND_TRANSFER = 'TRANSACTION_KIND_TRANSFER',
    TRANSACTION_KIND_EXCHANGE = 'TRANSACTION_KIND_EXCHANGE',
    TRANSACTION_KIND_FEE = 'TRANSACTION_KIND_FEE',
    TRANSACTION_KIND_TRADE = 'TRANSACTION_KIND_TRADE',
    TRANSACTION_KIND_TAX = 'TRANSACTION_KIND_TAX',
}