// src/lib/api/generated/models/v1TransactionStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1TransactionStatus {
    TRANSACTION_STATUS_UNSPECIFIED = 'TRANSACTION_STATUS_UNSPECIFIED',
    TRANSACTION_STATUS_REALIZED = 'TRANSACTION_STATUS_REALIZED',
    TRANSACTION_STATUS_REJECTED = 'TRANSACTION_STATUS_REJECTED',
    TRANSACTION_STATUS_PROCESSING = 'TRANSACTION_STATUS_PROCESSING',
}