// src/lib/api/generated/models/v1OptionType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1OptionType {
    OPTION_TYPE_UNSPECIFIED = 'OPTION_TYPE_UNSPECIFIED',
    OPTION_TYPE_CALL = 'OPTION_TYPE_CALL',
    OPTION_TYPE_PUT = 'OPTION_TYPE_PUT',
}